feat: Add .5 and 1.5 bathroom options to min and max dropdown in Add … (#1113)

+++ backend/core/src/units-summary/entities/unit-group.entity.ts
@@ -82,13 +82,13 @@ export class UnitGroup {
   @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
   floorMax?: number | null
 
-  @Column({ nullable: true, type: "integer" })
+  @Column({ nullable: true, type: "numeric" })
   @Expose()
   @IsOptional({ groups: [ValidationsGroupsEnum.default] })
   @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
   sqFeetMin?: number | null
 
-  @Column({ nullable: true, type: "integer" })
+  @Column({ nullable: true, type: "numeric" })
   @Expose()
   @IsOptional({ groups: [ValidationsGroupsEnum.default] })
   @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
@@ -113,13 +113,13 @@ export class UnitGroup {
   @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
   totalAvailable?: number | null
 
-  @Column({ nullable: true, type: "integer" })
+  @Column({ nullable: true, type: "numeric" })
   @Expose()
   @IsOptional({ groups: [ValidationsGroupsEnum.default] })
   @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
   bathroomMin?: number | null
 
-  @Column({ nullable: true, type: "integer" })
+  @Column({ nullable: true, type: "numeric" })
   @Expose()
   @IsOptional({ groups: [ValidationsGroupsEnum.default] })
   @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
